requirements.txt
python-dateutil==2.7.3
six==1.11.0
Flask==1.1.1
psycopg2==2.8.4
flask-sqlalchemy==2.4.1